+++ frontend/aid-vault/src/lib/AddDetails.svelte
