+++ frontend/aid-vault/src/lib/ApiSetDetails.svelte
