+++ frontend/aid-vault/src/lib/Counter.svelte
