+++ frontend/aid-vault/src/lib/Loader.svelte
