+++ frontend/aid-vault/src/lib/Navbar.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import "../../../app.css"
+    import "../app.css"
 
     export let enable : boolean;
